Move all views to testing_views
