+++ vue/sbc-common-components/src/router.ts
@@ -1,24 +1,24 @@
-import AuthMenuView from './views/AuthMenuView.vue'
-import AuthenticationOptionsDialogView from './views/AuthenticationOptionsDialogView.vue'
-import AuthenticationOptionsView from './views/AuthenticationOptionsView.vue'
-import BaseAddressView from './views/BaseAddressView.vue'
-import BrowserVersionAlertView from './views/BrowserVersionAlertView.vue'
-import FeeSummaryView from './views/FeeSummaryView.vue'
-import Home from './views/Home.vue'
-import LoaderView from './views/LoaderView.vue'
-import LoadingScreenView from './views/LoadingScreenView.vue'
-import LoginView from './views/LoginView.vue'
-import MobileDeviceAlertView from './views/MobileDeviceAlertView.vue'
-import NavigationBarView from './views/NavigationBarView.vue'
-import NotificationPanelView from './views/NotificationPanelView.vue'
-import PaySystemAlertView from './views/PaySystemAlertView.vue'
-import ProductSelectorView from './views/ProductSelectorView.vue'
-import SignIn from './views/SigninView.vue'
-import SignoutView from './views/SignoutView.vue'
-import SystemAlertView from './views/SystemAlertView.vue'
-import SystemBannerView from './views/SystemBannerView.vue'
-import SystemErrorModalView from './views/SystemErrorModalView.vue'
-import SystemErrorView from './views/SystemErrorView.vue'
+import AuthMenuView from './testing_views/AuthMenuView.vue'
+import AuthenticationOptionsDialogView from './testing_views/AuthenticationOptionsDialogView.vue'
+import AuthenticationOptionsView from './testing_views/AuthenticationOptionsView.vue'
+import BaseAddressView from './testing_views/BaseAddressView.vue'
+import BrowserVersionAlertView from './testing_views/BrowserVersionAlertView.vue'
+import FeeSummaryView from './testing_views/FeeSummaryView.vue'
+import Home from './testing_views/Home.vue'
+import LoaderView from './testing_views/LoaderView.vue'
+import LoadingScreenView from './testing_views/LoadingScreenView.vue'
+import LoginView from './testing_views/LoginView.vue'
+import MobileDeviceAlertView from './testing_views/MobileDeviceAlertView.vue'
+import NavigationBarView from './testing_views/NavigationBarView.vue'
+import NotificationPanelView from './testing_views/NotificationPanelView.vue'
+import PaySystemAlertView from './testing_views/PaySystemAlertView.vue'
+import ProductSelectorView from './testing_views/ProductSelectorView.vue'
+import SignIn from './testing_views/SigninView.vue'
+import SignoutView from './testing_views/SignoutView.vue'
+import SystemAlertView from './testing_views/SystemAlertView.vue'
+import SystemBannerView from './testing_views/SystemBannerView.vue'
+import SystemErrorModalView from './testing_views/SystemErrorModalView.vue'
+import SystemErrorView from './testing_views/SystemErrorView.vue'
 import Vue from 'vue'
 import Router from 'vue-router'
