+++ vue/sbc-common-components/src/testing_views/AuthMenuView.vue
