+++ vue/sbc-common-components/src/testing_views/AuthenticationOptionsDialogView.vue
