+++ vue/sbc-common-components/src/testing_views/AuthenticationOptionsView.vue
