+++ vue/sbc-common-components/src/testing_views/BaseAddressView.vue
