+++ vue/sbc-common-components/src/testing_views/BrowserVersionAlertView.vue
