+++ vue/sbc-common-components/src/testing_views/LoaderView.vue
