+++ vue/sbc-common-components/src/testing_views/LoadingScreenView.vue
