+++ vue/sbc-common-components/src/testing_views/LoginView.vue
