+++ vue/sbc-common-components/src/testing_views/MobileDeviceAlertView.vue
