+++ vue/sbc-common-components/src/testing_views/NavigationBarView.vue
