+++ vue/sbc-common-components/src/testing_views/NotificationPanelView.vue
