+++ vue/sbc-common-components/src/testing_views/PaySystemAlertView.vue
