+++ vue/sbc-common-components/src/testing_views/ProductSelectorView.vue
