+++ vue/sbc-common-components/src/testing_views/SigninView.vue
