+++ vue/sbc-common-components/src/testing_views/SignoutView.vue
