+++ vue/sbc-common-components/src/testing_views/SystemAlertView.vue
