+++ vue/sbc-common-components/src/testing_views/SystemBannerView.vue
