+++ vue/sbc-common-components/src/testing_views/SystemErrorModalView.vue
